fetch data from firebase partially competed
