+++ Data/data.ts
@@ -1,0 +1,278 @@
+const doctors = [
+  {
+    id: "D001",
+    name: "Dr. Sarah Johnson",
+    speciality: ["Cardiology"],
+    timings: "9:00 AM - 5:00 PM",
+    room_no: "203",
+    available: true,
+  },
+  {
+    id: "D002",
+    name: "Dr. Michael Brown",
+    speciality: ["General Medicine", "Family Practice"],
+    timings: "8:00 AM - 4:00 PM",
+    room_no: "101",
+    available: true,
+  },
+  {
+    id: "D003",
+    name: "Dr. Jessica Martinez",
+    speciality: ["Surgery", "Orthopedics"],
+    timings: "10:00 AM - 6:00 PM",
+    room_no: "305",
+    available: false,
+  },
+  {
+    id: "D004",
+    name: "Dr. David Lee",
+    speciality: ["Pulmonology", "Critical Care"],
+    timings: "7:00 AM - 7:00 PM",
+    room_no: "ICU-2",
+    available: true,
+  },
+  {
+    id: "D005",
+    name: "Dr. Patricia Wilson",
+    speciality: ["Emergency Medicine", "Trauma"],
+    timings: "Night Shift: 8:00 PM - 8:00 AM",
+    room_no: "ER-1",
+    available: true,
+  },
+  {
+    id: "D006",
+    name: "Dr. Elizabeth Chen",
+    speciality: ["Obstetrics", "Gynecology"],
+    timings: "9:00 AM - 3:00 PM",
+    room_no: "204",
+    available: false,
+  },
+  {
+    id: "D007",
+    name: "Dr. Robert Taylor",
+    speciality: ["Cardiothoracic Surgery"],
+    timings: "8:00 AM - 4:00 PM",
+    room_no: "OR-3",
+    available: true,
+  },
+  {
+    id: "D008",
+    name: "Dr. James Wilson",
+    speciality: ["Neurology", "Stroke Care"],
+    timings: "10:00 AM - 6:00 PM",
+    room_no: "302",
+    available: true,
+  },
+];
+
+// patients //////////////////////////////////////////////
+const patients = [
+  {
+    id: "P001",
+    name: "John Smith",
+    issues: ["Chest Pain", "High Blood Pressure"],
+    type: "Emergency",
+    date: new Date(2023, 3, 21),
+    ipd_no: "IPD2023042101",
+    doctor: "Dr. Sarah Johnson",
+    staff: ["Nurse Amy", "Tech Rob"],
+    medicines: ["Aspirin", "Lisinopril"],
+  },
+  {
+    id: "P002",
+    name: "Emily Davis",
+    issues: ["Annual Checkup"],
+    type: "Regular",
+    date: new Date(2023, 3, 20),
+    ipd_no: "IPD2023042001",
+    doctor: "Dr. Michael Brown",
+    staff: ["Nurse Tom"],
+    medicines: [],
+  },
+  {
+    id: "P003",
+    name: "Robert Wilson",
+    issues: ["Appendicitis"],
+    type: "Surgery",
+    date: new Date(2023, 3, 19),
+    ipd_no: "IPD2023041901",
+    doctor: "Dr. Jessica Martinez",
+    staff: ["Nurse Emma", "Nurse Dave", "Tech Samantha"],
+    medicines: ["Antibiotics", "Painkillers", "IV Fluids"],
+  },
+  {
+    id: "P004",
+    name: "Lisa Thompson",
+    issues: ["Respiratory Failure", "Pneumonia"],
+    type: "ICU",
+    date: new Date(2023, 3, 18),
+    ipd_no: "IPD2023041801",
+    doctor: "Dr. David Lee",
+    staff: ["Nurse Alex", "Nurse Maria", "Tech James"],
+    medicines: ["Ventolin", "Antibiotics", "Steroids", "Oxygen Therapy"],
+  },
+  {
+    id: "P005",
+    name: "Michael Johnson",
+    issues: ["Broken Arm", "Concussion"],
+    type: "Emergency",
+    date: new Date(2023, 3, 17),
+    ipd_no: "IPD2023041701",
+    doctor: "Dr. Patricia Wilson",
+    staff: ["Nurse Bob", "Tech Lucy"],
+    medicines: ["Painkillers", "Anti-inflammatory"],
+  },
+  {
+    id: "P006",
+    name: "Jennifer Brown",
+    issues: ["Pregnancy Checkup"],
+    type: "Regular",
+    date: new Date(2023, 3, 16),
+    ipd_no: "IPD2023041601",
+    doctor: "Dr. Elizabeth Chen",
+    staff: ["Nurse Kelly"],
+    medicines: ["Prenatal Vitamins"],
+  },
+  {
+    id: "P007",
+    name: "David Miller",
+    issues: ["Heart Surgery", "Coronary Artery Disease"],
+    type: "Surgery",
+    date: new Date(2023, 3, 15),
+    ipd_no: "IPD2023041501",
+    doctor: "Dr. Robert Taylor",
+    staff: ["Nurse Sarah", "Nurse John", "Tech Diana", "Tech George"],
+    medicines: ["Blood Thinners", "Beta Blockers", "Pain Medication", "Antibiotics"],
+  },
+  {
+    id: "P008",
+    name: "Susan Anderson",
+    issues: ["Stroke", "Hypertension"],
+    type: "ICU",
+    date: new Date(2023, 3, 14),
+    ipd_no: "IPD2023041401",
+    doctor: "Dr. James Wilson",
+    staff: ["Nurse Emily", "Nurse Michael", "Tech Robert"],
+    medicines: ["Blood Thinners", "Anti-hypertensives", "Statins"],
+  },
+];
+
+
+// Sample data
+const supplies = [
+  {
+    item_id: 1001,
+    name: "Disposable Gloves",
+    type: "PPE",
+    delivery_date: new Date(2023, 2, 15),
+    expiry_date: new Date(2025, 2, 15),
+    expired: false,
+  },
+  {
+    item_id: 1002,
+    name: "Surgical Masks",
+    type: "PPE",
+    delivery_date: new Date(2023, 1, 20),
+    expiry_date: new Date(2025, 1, 20),
+    expired: false,
+  },
+  {
+    item_id: 1003,
+    name: "Antibiotics - Amoxicillin",
+    type: "Medication",
+    delivery_date: new Date(2023, 0, 10),
+    expiry_date: new Date(2023, 6, 10),
+    expired: true,
+  },
+  {
+    item_id: 1004,
+    name: "IV Solution - Normal Saline",
+    type: "Medical Supply",
+    delivery_date: new Date(2023, 3, 5),
+    expiry_date: new Date(2025, 3, 5),
+    expired: false,
+  },
+  {
+    item_id: 1005,
+    name: "Syringes 10ml",
+    type: "Medical Supply",
+    delivery_date: new Date(2023, 2, 25),
+    expiry_date: new Date(2026, 2, 25),
+    expired: false,
+  },
+  {
+    item_id: 1006,
+    name: "Bandages",
+    type: "Medical Supply",
+    delivery_date: new Date(2023, 4, 1),
+    expiry_date: new Date(2026, 4, 1),
+    expired: false,
+  },
+  {
+    item_id: 1007,
+    name: "Antiseptic Solution",
+    type: "Medical Supply",
+    delivery_date: new Date(2023, 1, 15),
+    expiry_date: new Date(2024, 1, 15),
+    expired: false,
+  },
+  {
+    item_id: 1008,
+    name: "Painkillers - Ibuprofen",
+    type: "Medication",
+    delivery_date: new Date(2023, 0, 20),
+    expiry_date: new Date(2023, 7, 20),
+    expired: true,
+  },
+];
+
+const bloodInventory = [
+  {
+    item_id: 5001,
+    blood_group: "A+",
+    type: "Whole Blood",
+    delivery_date: new Date(2023, 3, 20),
+    expiry_date: new Date(2023, 4, 18),
+    expired: false,
+  },
+  {
+    item_id: 5002,
+    blood_group: "B+",
+    type: "Platelets",
+    delivery_date: new Date(2023, 3, 22),
+    expiry_date: new Date(2023, 3, 29),
+    expired: true,
+  },
+  {
+    item_id: 5003,
+    blood_group: "O-",
+    type: "Whole Blood",
+    delivery_date: new Date(2023, 3, 25),
+    expiry_date: new Date(2023, 4, 23),
+    expired: false,
+  },
+  {
+    item_id: 5004,
+    blood_group: "AB+",
+    type: "Plasma",
+    delivery_date: new Date(2023, 3, 26),
+    expiry_date: new Date(2023, 6, 26),
+    expired: false,
+  },
+  {
+    item_id: 5005,
+    blood_group: "A-",
+    type: "Red Blood Cells",
+    delivery_date: new Date(2023, 3, 27),
+    expiry_date: new Date(2023, 5, 27),
+    expired: false,
+  },
+  {
+    item_id: 5006,
+    blood_group: "O+",
+    type: "Whole Blood",
+    delivery_date: new Date(2023, 3, 15),
+    expiry_date: new Date(2023, 4, 13),
+    expired: false,
+  },
+];
